% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFPlus-utils.R
\name{dummy_code}
\alias{dummy_code}
\title{Dummy code factor variables in x and x_test.}
\usage{
dummy_code(x, x_test = NULL)
}
\arguments{
\item{x}{A data frame containing the training data.}

\item{x_test}{A data frame containing the test data.}
}
\description{
Dummy code factor variables in x and x_test.
}
