% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFPlus-utils.R
\name{extract_psi_fast}
\alias{extract_psi_fast}
\title{Extract Psi (i.e., stump) matrix from tree.}
\usage{
extract_psi_fast(
  x,
  tree_info,
  tree_paths,
  node_preds,
  psi_train = NULL,
  inbag_counts = NULL,
  normalize = FALSE,
  unordered_factors = NULL
)
}
\arguments{
\item{x}{Data matrix for which to compute Psi.}

\item{tree_info}{Output of \code{\link[ranger:treeInfo]{ranger::treeInfo()}} for a single tree.}

\item{tree_paths}{Output of \code{\link[=get_tree_paths_fast]{get_tree_paths_fast()}}.}

\item{node_preds}{Terminal node predictions corresponding to x.}

\item{psi_train}{(Optional) Psi matrix for training data. If provided, the
output will be normalized using the same normalizing constants as the
training data. Otherwise, the output will be normalized using the
normalizing constants computed from the input data. Only used if
\code{normalize = TRUE}.}

\item{inbag_counts}{(Optional) Number of times each training observation was
sampled in the tree (from RF+). Only used if \code{normalize = TRUE}.}

\item{normalize}{Whether to normalize the stumps so that larger stumps
have larger magnitudes.}

\item{unordered_factors}{Names of unordered factors in x.}
}
\description{
Extract Psi (i.e., stump) matrix from tree.
}
