% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFPlus-utils.R
\name{get_augmented_x}
\alias{get_augmented_x}
\title{Get augmented x matrix (i.e., (x_s, psi)) from x and psi.}
\usage{
get_augmented_x(
  x,
  psi,
  tree_info = NULL,
  include_raw = TRUE,
  normalize = FALSE,
  normalize_factor = NULL
)
}
\arguments{
\item{x}{A data frame or matrix of features.}

\item{psi}{A data frame or matrix of psi values.}

\item{tree_info}{Output of \code{\link[ranger:treeInfo]{ranger::treeInfo()}} for a single tree.}

\item{include_raw}{Whether to include the raw features in the RF+ estimator.}

\item{normalize}{Whether to normalize psi values so that largest stump per
feature has the same standard deviation as its raw feature.}

\item{normalize_factor}{A numeric vector of length equal to the number of
features in x. If provided, this vector will be used to normalize the
features in x. If NULL, the standard deviation of each feature in x will
be used.}
}
\description{
Get augmented x matrix (i.e., (x_s, psi)) from x and psi.
}
