% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance.R
\name{get_feature_importances}
\alias{get_feature_importances}
\title{Compute feature importance for RF+ estimators.}
\usage{
get_feature_importances(
  object,
  x,
  y,
  type = c("permute", "mdi+"),
  B = 10,
  local = FALSE,
  use_rcpp = TRUE
)
}
\arguments{
\item{object}{An object of class \code{rfplus}.}

\item{x}{A matrix or data frame of predictors.}

\item{y}{A vector of responses.}

\item{type}{The type of feature importance to compute. Can be one of
\code{"permute"} or \code{"mdi+"}. See \code{\link{permute_fi}} and
\code{\link{mdiplus_fi}} for details.}

\item{B}{The number of permutations to perform. If \code{B} is a list of
vectors of indices, then the indices will be used to permute the features.}

\item{local}{Whether to compute local feature importance.}

\item{use_rcpp}{Whether to use Rcpp. Should remove this argument before
release. Keep only for debugging.}
}
\description{
Compute feature importance for RF+ estimators.
}
