% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFPlus-utils.R
\name{get_forest_paths}
\alias{get_forest_paths}
\title{Extract all root-to-leaf paths in a forest (non-Rcpp version).}
\usage{
get_forest_paths(tree_infos)
}
\arguments{
\item{tree_infos}{List of size num.trees with each entry being the output of
\code{\link[ranger:treeInfo]{ranger::treeInfo()}}.}
}
\description{
Extract all root-to-leaf paths in a forest (non-Rcpp version).
}
