% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance.R
\name{get_grouped_tree_features}
\alias{get_grouped_tree_features}
\title{Get stump features corresponding to each raw feature}
\usage{
get_grouped_tree_features(orig_colnames, aug_colnames, tree_info)
}
\arguments{
\item{orig_colnames}{Original (raw) column names.}

\item{aug_colnames}{Augmented (raw + stump) column names.}

\item{tree_info}{Output of \code{\link[ranger:treeInfo]{ranger::treeInfo()}} for a single tree.}
}
\description{
Get stump features corresponding to each raw feature
}
