% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFPlus-utils.R
\name{get_train_test_psis}
\alias{get_train_test_psis}
\title{Wrapper to get Psi for training and test data.}
\usage{
get_train_test_psis(object, x, use_rcpp = TRUE)
}
\arguments{
\item{object}{An object of class \code{rfplus}.}

\item{x}{A data frame containing the test data.}

\item{use_rcpp}{A logical indicating whether to use Rcpp code.}
}
\description{
Wrapper to get Psi for training and test data.
}
