% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFPlus-utils.R
\name{get_tree_paths}
\alias{get_tree_paths}
\title{Extract all root-to-leaf paths in a tree (non-Rcpp version).}
\usage{
get_tree_paths(tree_info)
}
\arguments{
\item{tree_info}{Output of \code{\link[ranger:treeInfo]{ranger::treeInfo()}} for a single tree.}
}
\description{
Extract all root-to-leaf paths in a tree (non-Rcpp version).
}
