% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFPlus-utils.R
\name{get_x_psi}
\alias{get_x_psi}
\title{Preprocess x before computing Psi matrix.}
\usage{
get_x_psi(x, x_train, fit)
}
\arguments{
\item{x}{A data frame containing the data to preprocess.}

\item{x_train}{A data frame containing the training data used for fitting
the RF+ model.}

\item{fit}{An object of class \code{rfplus}.}
}
\description{
Preprocess x before computing Psi matrix.
}
