% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance.R
\name{mdiplus_fi}
\alias{mdiplus_fi}
\title{Compute MDI+ importance for RF+ estimators.}
\usage{
mdiplus_fi(
  object,
  x,
  y,
  col_means = NULL,
  metric = rsq_vec_narm,
  grouped_features = NULL,
  local = FALSE,
  local_metric = yardstick::rmse_vec
)
}
\arguments{
\item{object}{An object of class \code{rfplus}.}

\item{x}{A matrix or data frame of predictors.}

\item{y}{A vector of responses.}

\item{col_means}{A vector of column means to use for the global feature
importance. If \code{NULL}, then the column means of \code{x} will be used.}

\item{metric}{A function that takes two vectors and returns a scalar.}

\item{grouped_features}{A list of vectors of feature names. Each vector
corresponds to a group of features that should be permuted together.}

\item{local}{Whether to compute local feature importance.}

\item{local_metric}{A function that takes two vectors and returns a vector
of the same length.}
}
\description{
Compute MDI+ importance for RF+ estimators.
}
