% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance.R
\name{permute_fi}
\alias{permute_fi}
\title{Compute permutation importance for RF+ estimators.}
\usage{
permute_fi(
  object,
  x,
  y,
  metric = yardstick::rmse_vec,
  grouped_features = NULL,
  B = 10,
  local = FALSE,
  local_metric = metric,
  return_preds = FALSE
)
}
\arguments{
\item{object}{An object of class \code{rfplus}.}

\item{x}{A matrix or data frame of predictors.}

\item{y}{A vector of responses.}

\item{metric}{A function that takes two vectors and returns a scalar.}

\item{grouped_features}{A list of vectors of feature names. Each vector
corresponds to a group of features that should be permuted together.}

\item{B}{The number of permutations to perform. If \code{B} is a list of
vectors of indices, then the indices will be used to permute the features.}

\item{local}{Whether to compute local feature importance.}

\item{local_metric}{A function that takes two vectors and returns a vector
of the same length.}

\item{return_preds}{Whether to return the predictions for each permutation.}
}
\description{
Compute permutation importance for RF+ estimators.
}
