% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFPlus.R
\name{predict.rfplus}
\alias{predict.rfplus}
\title{Predict using Random Forest Plus (RF+)}
\usage{
\method{predict}{rfplus}(object, x, type = c("response"), return_all = FALSE, use_rcpp = TRUE)
}
\arguments{
\item{object}{An object of class \code{rfplus}.}

\item{x}{A matrix or data frame of observations.}

\item{type}{The type of prediction to return.}

\item{return_all}{Whether to return predictions from all trees (rather than
the average prediction).}

\item{use_rcpp}{Whether to use the Rcpp. Should remove this argument before
release. Keep only for debugging.}
}
\description{
Predict using Random Forest Plus (RF+)
}
