% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFPlus.R
\name{rfplus_cv}
\alias{rfplus_cv}
\title{Random Forest Plus (RF+) with cross-validation wrapepr.}
\usage{
rfplus_cv(
  x,
  y,
  include_raw = TRUE,
  normalize_stumps = FALSE,
  normalize_raw = FALSE,
  family = c("linear", "logistic"),
  sample_split = c("none", "oob"),
  ntrees = 500,
  ntrees_cv = ntrees,
  mtry = NULL,
  num.threads = 1,
  lambda_xs = 0,
  lambda_ts = NULL,
  cv = 5,
  parallel = FALSE,
  low_mem = TRUE,
  use_rcpp = TRUE,
  ...
)
}
\arguments{
\item{x}{A matrix or data frame of predictors.}

\item{y}{A vector of responses.}

\item{include_raw}{Whether to include the raw features in the RF+ estimator.}

\item{normalize_stumps}{Whether to normalize the stumps in the RF+ estimator.}

\item{normalize_raw}{Whether to normalize the raw features in the RF+
estimator.}

\item{family}{The family of the response. Either "linear" or "logistic".}

\item{sample_split}{How to perform sample splitting for the RF+ estimator.
Must be one of "none" or "oob".}

\item{ntrees}{The number of trees to grow.}

\item{ntrees_cv}{Number of trees to use for cross-validation. Defaults to
\code{ntrees}. If \code{ntrees_cv} is less than \code{ntrees}, then
\code{ntrees_cv} trees will be tuned via cross-validation and the remaining
\code{ntrees - ntrees_cv} trees will choose a hyperparameter set randomly
from the best hyperparameters from the cross-validated \code{ntrees_cv}
trees.}

\item{mtry}{The number of variables to sample as candidates at each split.}

\item{num.threads}{The number of threads to use for training RF+.}

\item{lambda_xs}{Vector or list of regularization parameters for the raw
features to search over.}

\item{lambda_ts}{Vector or list of regularization parameters for the stump
features to search over.}

\item{cv}{Number of folds for cross-validation.}

\item{parallel}{Whether to use parallel processing.}

\item{low_mem}{Whether to use low memory mode. Should remove this argument
before release. Keep only for debugging.}

\item{use_rcpp}{Whether to use Rcpp. Should remove this argument before
release. Keep only for debugging.}

\item{...}{Other arguments passed to \code{ranger::ranger}.}
}
\description{
Random Forest Plus (RF+) with cross-validation wrapepr.
}
