% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFPlus-utils.R
\name{rsq_vec_narm}
\alias{rsq_vec_narm}
\title{Compute R-squared between two vectors while returning 0 if the estimate is
constant.}
\usage{
rsq_vec_narm(truth, estimate, ...)
}
\arguments{
\item{truth}{The column identifier for the true results
(that is \code{numeric}). This should be an unquoted column name although
this argument is passed by expression and supports
\link[rlang:topic-inject]{quasiquotation} (you can unquote column
names). For \verb{_vec()} functions, a \code{numeric} vector.}

\item{estimate}{The column identifier for the predicted
results (that is also \code{numeric}). As with \code{truth} this can be
specified different ways but the primary method is to use an
unquoted variable name. For \verb{_vec()} functions, a \code{numeric} vector.}

\item{...}{Not currently used.}
}
\description{
Compute R-squared between two vectors while returning 0 if the estimate is
constant.
}
